# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management System API",
        "description": "Hello my name is Tresten Pool aspiring backend developer. I created this api for my advanced software engineering class",
        "contact": {
            "email": "trestenpool@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://ec2-54-90-227-153.compute-1.amazonaws.com/api/"
        }
    ],
    "paths": {
...
        "/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all the products",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Store a new product",
                "operationId": "storeNewProduct",
                "requestBody": {
                    "description": "Pass in the product name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "409": {
                        "description": "Conflict, there is already a product with that name"
                    }
                }
            }
        },
        "/product/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by product id",
                "operationId": "GetProductID",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "The product id route parameter is not valid. Must be numeric.."
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass in the product name",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "default": ""
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid request body"
                    },
                    "404": {
                        "description": "Product with the id provided was not found"
                    },
                    "409": {
                        "description": "Conflict, there is already a product with that name"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product by it's ID",
                "operationId": "DeleteProduct",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        }
    }
}